General improvements, fixes and updates
Next up:
- Tidy up code.
- Remove duplicate code and making appropriate functions for said duplicated code.
- Make code of conduct, contributing guidelines and issue- and pull request templates.
- See what features users want in a bot, it is hard for me to know what to make.
- Remove command usage stats since I do not want to infringe on people's privacy.
- See if we really need an API and web dashboard, otherwise just remove it.
- Add commands utilising the Tetr.io API. (?)
- Make it easier to extend Wavy's feature-set. (?)
- See if Pycord is still the best API wrapper to use, Discord.py is back but it lacks the new features Discord has added, but in my experience it has always worked much better than Pycord. I am not against rewriting the bot again (nor doing that in another language), but I feel like people will get sick and tired of me constantly rewriting everything.

The above are just ideas, and will most likely get their own respective commits so it doesn't look like the project is dead for days on end. I have not decided on the idea(s) with a question mark `?` yet.

If you want a new feature, or you want to provide feedback, please do! I have no idea what to make next, and it's honestly quite saddening to see a project I worked on just die because I have no idea what to do with it. (But yeah people also didn't ask me to make a bot, so I shouldn't expect much honestly.) I hope to be able to attract more people to the bot, but right now I am basically in an infinite loop of trying to get feedback and such and not having enough people to ask for feedback.

I should probably just make a development ramblings channel on Discord for this stuff, but oh well. Thanks for reading and have a great day! (^-^)

~Robert